cast to appropriate type

+++ qsplit/src/org/qcmg/split/SingleLevelSplit.java
@@ -494,7 +494,7 @@ public class SingleLevelSplit extends Split {
 				throw new Exception("Input file contains records lacking ZC integer attribute");
 			}
 			
-			SAMFileWriter writer = zcToWriterMap.get(zc);
+			SAMFileWriter writer = zcToWriterMap.get((Integer)zc);
 
 			if (null == writer) {
 				closeWriters();
